requirements.txt
Flask==1.1.2
gunicorn==20.0.4
psycopg2-binary==2.8.6